requirements.txt
httpx>=0.27,<0.28
pandas>=2.2,<2.3
python-dateutil>=2.8,<3.0
supabase>=2.0,<3.0
python-dotenv>=1.0,<2.0
pydantic>=2.7,<3.0
pydantic-settings>=2.6,<3.0
